+++ Block/Adminhtml/Form/Field/MethodStatusDisplay/Standard.php
@@ -1,0 +1,16 @@
+<?php
+
+declare(strict_types=1);
+
+namespace ETechFlow\NextDayEligibility\Block\Adminhtml\Form\Field\MethodStatusDisplay;
+
+use ETechFlow\NextDayEligibility\Block\Adminhtml\Form\Field\MethodStatusDisplay;
+use ETechFlow\NextDayEligibility\Model\ShippingMethodAvailability;
+
+class Standard extends MethodStatusDisplay
+{
+    protected function getType(): string
+    {
+        return ShippingMethodAvailability::TYPE_STANDARD;
+    }
+}